rename mip-fh-bus
