+++ mip-fh-bus/README.md
@@ -1,23 +1,23 @@
-# mip-fh-sex 
+# mip-fh-bus 
 
-mip-fh-sex 用于改造两性详情页面的脚本代码
+mip-fh-bus 用于改造两性详情页面的脚本代码
 
 |描述|脚本mip改造|
 |---|---|
 |类型|脚本|
 |支持布局|N/S|
-|所需脚本|https://mipcache.bdstatic.com/static/v1/mip-fh-sex/mip-fh-sex.js|
+|所需脚本|https://mipcache.bdstatic.com/static/v1/mip-fh-bus/mip-fh-bus.js|
 
 ## 示例
 
 在MIP HTML中,直接使用标签, 只需使用一次。示例如下:
 
 ```html
-<div id="FhwapContent">
+<div class="FhwapContent">
  <p>
     <mip-img src="http://sex.fh21.com.cn/uploads/allimg/160323/2644_160323151044_1.jpg"/>
  </p>
 </div>
-<mip-fh-sex></mip-fh-sex>
+<mip-fh-bus></mip-fh-bus>
 ```
