+++ mip-fh-bus/bus-set-img-ad.js
@@ -41,7 +41,7 @@ define(function (require) {
             $img.after(node);
         };
 
-        this.loadScript('http://1.feihua.com/au3a1ecf91ffc8f038db4c3e8da4f73ffa54acde0b36.js',function(){
+        this.loadScript('http://1.feihua.com/au3a1ecf91ffc8f038db4c3e8da4f73ffa54acde0b36.js', function () {
             var timer = setTimeout(function () {
                 document.write = docw;
                 clearTimeout(timer);
