+++ mip-fh-bus/package.json
@@ -1,5 +1,5 @@
 {
-    "name": "mip-fh-sex",
+    "name": "mip-fh-bus",
     "version": "1.0.0",
     "author": {
         "name": "laoono",
